Modification of SelectChat and GroupChat class
Modified the SelectUsers class and GroupChat class. After the user selects which users they want. add a new node in every participating group members user node in the database, then UI pass to the GroupChat.  Then, it is OK to submit a message to the groupid, contains user name and message. this message will be going under the messages section in our data base so messages->groupid->msgs. It is Ok to compile and run in the project. I also add an additional member in UserDetails.